Routing for changing any of the location fields
If changing any aspect of the location from the review location page it needs to step through each location page after that one and then to review location.
If changing any aspect of the location from check answers at the end of the journey it needs to step through each location page after that one and then to review location then jump to check answers.
In other words always review the location when making any changes to it.